features added, big progress
Made it so that when you hover over a dot in turns either red or green. If its a legal move you can do it turns green, if not it turns red. when you click a green dot, it will draw a line that connects your current dot in to the next dot