Implemented cross out of items in pipes (2/4) and myTurn bool
I coded the ability to be able to hover and click over the items in the pipe and have it show a circle with an (x) inside. If you hover over it the red x shows, and if you move away it goes away, when you click it, it stays. If its not your turn you are not able to click it and make it stay. I added this boolean myTurn to determine if its the players turn, however since we don't have network features yet, I have placed a hidden box that shows up when you hover over it in the top left, if you click on it, the boolean value becomes true.